requirements.txt
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
google-cloud-firestore>=2.11.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pydantic>=2.0.0
python-dateutil>=2.8.2
google-adk
flask==2.3.3
gunicorn==21.2.0
streamlit==1.28.0
firebase-admin>=6.2.0